Remove supabase admin
