+++ prisma/schema/auth.prisma
@@ -8,6 +8,9 @@ model User {
     image   String?
     isAdmin Boolean @default(false)
 
+    billingAddress Json?
+    paymentMethod  Json?
+
     emailVerified DateTime?
     passwordHash  String?
