+++ prisma/schema/stripe.prisma
@@ -55,7 +55,7 @@ model Product {
 }
 
 model ProductAmount {
-    id        String   @id @default(cuid())
+    id        String   @id
     createdAt DateTime @default(now())
     updatedAt DateTime @updatedAt
 
@@ -98,14 +98,15 @@ model Subscription {
     userId             String
     status             SubscriptionStatus?
     quantity           Int?
+    metadata           Json?
     cancelAtPeriodEnd  Boolean?
     currentPeriodStart DateTime            @default(now())
     currentPeriodEnd   DateTime            @default(now())
-    endedAt            DateTime            @default(now())
-    cancelAt           DateTime            @default(now())
-    canceledAt         DateTime            @default(now())
-    trialStart         DateTime            @default(now())
-    trialEnd           DateTime            @default(now())
+    endedAt            DateTime?
+    cancelAt           DateTime?
+    canceledAt         DateTime?
+    trialStart         DateTime?
+    trialEnd           DateTime?
 
     price Price @relation(fields: [priceId], references: [id], onDelete: Cascade)
     user  User  @relation(fields: [userId], references: [id], onDelete: Cascade)
